+++ crates/nostr-relay-pool/src/pool/constants.rs
@@ -2,4 +2,9 @@
 // Copyright (c) 2023-2024 Rust Nostr Developers
 // Distributed under the MIT software license
 
+//! Constants
+
 pub(super) const MAX_CONNECTING_CHUNK: usize = 100;
+
+/// Relay Pool default notification channel size
+pub const DEFAULT_NOTIFICATION_CHANNEL_SIZE: usize = 4096;